requirements.txt
attrs==23.2.0
bidict==0.23.1
certifi==2024.7.4
charset-normalizer==3.3.2
docker==7.1.0
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
PySocks==1.7.1
python-engineio==4.12.3
python-socketio==5.14.2
requests==2.32.3
selenium==4.22.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
